requirements.txt
numpy
scipy
matplotlib
pandas
torch
torchvision
scikit-learn
lightkurve
batman-package
streamlit
shap
tqdm
seaborn
captum
astropyquantity